+++ src/components/EditionDetail/Edition.tsx
@@ -2,7 +2,6 @@ import React from "react";
 import { useRouter } from "next/router";
 import {
   Box,
-  Breadcrumbs,
   Card,
   CardActions,
   CardContent,
@@ -11,12 +10,6 @@ import {
   Heading,
   HorizontalRule,
   SimpleGrid,
-  Template,
-  TemplateBreakout,
-  TemplateContent,
-  TemplateContentPrimary,
-  TemplateContentTop,
-  TemplateFooter,
   Text,
   Toggle,
 } from "@nypl/design-system-react-components";
@@ -26,7 +19,6 @@ import { ApiEdition, EditionResult } from "~/src/types/EditionQuery";
 import EditionDetailDefinitionList from "~/src/components/EditionDetailDefinitionList/EditionDetailDefinitionList";
 import Link from "~/src/components/Link/Link";
 import { InstanceCard } from "../InstanceCard/InstanceCard";
-import { defaultBreadcrumbs } from "~/src/constants/labels";
 import SearchHeader from "../SearchHeader/SearchHeader";
 import {
   joinArrayOfElements,
@@ -36,6 +28,13 @@ import { MAX_TITLE_LENGTH } from "~/src/constants/editioncard";
 import { Instance } from "~/src/types/DataModel";
 import EditionCardUtils from "~/src/util/EditionCardUtils";
 import { PLACEHOLDER_LINK } from "~/src/constants/collection";
+import DrbTemplate, {
+  DrbBreakout,
+  DrbContent,
+  DrbContentPrimary,
+  DrbContentTop,
+  DrbHeader,
+} from "../DrbTemplate/DrbTemplate";
 
 const Edition: React.FC<{ editionResult: EditionResult; backUrl?: string }> = (
   props
@@ -78,23 +77,21 @@ const Edition: React.FC<{ editionResult: EditionResult; backUrl?: string }> = (
   };
 
   return (
-    <Template>
-      <TemplateBreakout>
-        <Breadcrumbs
-          breadcrumbsType="research"
-          breadcrumbsData={[
-            ...defaultBreadcrumbs,
-            {
-              url: `/work/${edition.work_uuid}`,
-              text: truncateStringOnWhitespace(edition.title, MAX_TITLE_LENGTH),
-            },
-          ]}
-        />
-        <SearchHeader />
-      </TemplateBreakout>
-
-      <TemplateContent>
-        <TemplateContentTop>
+    <DrbTemplate>
+      <DrbBreakout
+        breadcrumbsData={[
+          {
+            url: `/work/${edition.work_uuid}`,
+            text: truncateStringOnWhitespace(edition.title, MAX_TITLE_LENGTH),
+          },
+        ]}
+      >
+        <DrbHeader>
+          <SearchHeader />
+        </DrbHeader>
+      </DrbBreakout>
+      <DrbContent>
+        <DrbContentTop>
           <Flex direction={{ base: "column", md: "row" }}>
             {edition && (
               <Heading level="one">
@@ -169,8 +166,8 @@ const Edition: React.FC<{ editionResult: EditionResult; backUrl?: string }> = (
               </CardActions>
             </Card>
           )}
-        </TemplateContentTop>
-        <TemplateContentPrimary>
+        </DrbContentTop>
+        <DrbContentPrimary>
           <EditionDetailDefinitionList edition={edition} />
           <HorizontalRule bg="section.research.primary" />
           {edition.instances && (
@@ -197,10 +194,9 @@ const Edition: React.FC<{ editionResult: EditionResult; backUrl?: string }> = (
               />
             ))}
           </SimpleGrid>
-        </TemplateContentPrimary>
-      </TemplateContent>
-      <TemplateFooter />
-    </Template>
+        </DrbContentPrimary>
+      </DrbContent>
+    </DrbTemplate>
   );
 };