+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - Fix: Small adjustments to Playwright tests
 - Upgrade testing-library packages to v14 and jest to v29
+- Remove TemplateFooter components
+- Replace Template components with custom DrbTemplate component to reduce repeat components
+- Upgrade to NYPL Design System 1.7.3
 - Replace fireEvent with userEvent
 
 ## [0.17.4]